requirements.txt
asgiref==3.9.1
Django==5.2.4
djangorestframework==3.16.1
python-dotenv==1.1.1
sqlparse==0.5.3
